+++ NEWS.md
@@ -1,5 +1,6 @@
 # wrappedtools 0.9.3
 - bugfix in ColSeeker
+- ColSeeker allows vector of classes
 
 # wrappedtools 0.9.2
 - new function cat_desc_table to create summary tables